requirements-dev.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
fastapi==0.129.0
h11==0.16.0
idna==3.11
joblib==1.5.3
motor==3.7.1
nltk==3.9.2
owlready2==0.50
passlib==1.7.4
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
python-multipart==0.0.22
regex==2026.2.19
starlette==0.52.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
